# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Garuka API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analytics/overview": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Analytics Overview",
        "operationId": "get_analytics_overview",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsOverviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/trends": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Attendance Trends",
        "operationId": "get_attendance_trends",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 90,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyAttendanceTrendItem"
                  },
                  "title": "Response Get Attendance Trends"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/schools-compare": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Compare Schools",
        "operationId": "compare_schools",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SchoolCompareItem"
                  },
                  "type": "array",
                  "title": "Response Compare Schools"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActiveTermInfo": {
        "properties": {
          "academic_year": {
            "type": "integer",
            "title": "Academic Year"
          },
          "term_no": {
            "type": "integer",
            "title": "Term No"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "academic_year",
          "term_no",
          "name"
        ],
        "title": "ActiveTermInfo"
      },
      "AnalyticsOverviewResponse": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Africa/Kigali (UTC+2)"
          },
          "kigali_today": {
            "type": "string",
            "title": "Kigali Today"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeInfo"
          },
          "kpis": {
            "$ref": "#/components/schemas/OverviewKPIs"
          },
          "cases_by_level": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Cases By Level"
          },
          "active_term": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActiveTermInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "overdue_cases": {
            "items": {
              "$ref": "#/components/schemas/OverdueCaseItem"
            },
            "type": "array",
            "title": "Overdue Cases"
          },
          "reporting_gaps": {
            "items": {
              "$ref": "#/components/schemas/ReportingGapItem"
            },
            "type": "array",
            "title": "Reporting Gaps"
          },
          "classes_missing_today": {
            "items": {
              "$ref": "#/components/schemas/MissingClassItem"
            },
            "type": "array",
            "title": "Classes Missing Today"
          },
          "recent_escalations": {
            "items": {
              "$ref": "#/components/schemas/RecentEscalationItem"
            },
            "type": "array",
            "title": "Recent Escalations"
          },
          "pending_help_requests": {
            "items": {
              "$ref": "#/components/schemas/PendingHelpRequestItem"
            },
            "type": "array",
            "title": "Pending Help Requests"
          },
          "attendance_trend": {
            "items": {
              "$ref": "#/components/schemas/DailyAttendanceTrendItem"
            },
            "type": "array",
            "title": "Attendance Trend"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "kigali_today",
          "scope",
          "kpis"
        ],
        "title": "AnalyticsOverviewResponse"
      },
...
      "DailyAttendanceTrendItem": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "present": {
            "type": "integer",
            "title": "Present"
          },
          "absent": {
            "type": "integer",
            "title": "Absent"
          },
          "enrolled": {
            "type": "integer",
            "title": "Enrolled"
          },
          "cases": {
            "type": "integer",
            "title": "Cases",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "date",
          "present",
          "absent",
          "enrolled"
        ],
        "title": "DailyAttendanceTrendItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissingClassItem": {
        "properties": {
          "class_id": {
            "type": "string",
            "format": "uuid",
            "title": "Class Id"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          },
          "school_name": {
            "type": "string",
            "title": "School Name"
          }
        },
        "type": "object",
        "required": [
          "class_id",
          "class_name",
          "school_name"
        ],
        "title": "MissingClassItem"
      },
      "OverdueCaseItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "school_name": {
            "type": "string",
            "title": "School Name"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "days_open": {
            "type": "integer",
            "title": "Days Open"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ref",
          "student_name",
          "school_name",
          "level",
          "opened_at",
          "days_open"
        ],
        "title": "OverdueCaseItem"
      },
      "OverviewKPIs": {
        "properties": {
          "students_active": {
            "type": "integer",
            "title": "Students Active",
            "default": 0
          },
          "absent_today": {
            "type": "integer",
            "title": "Absent Today",
            "default": 0
          },
          "open_cases": {
            "type": "integer",
            "title": "Open Cases",
            "default": 0
          },
          "new_cases_7d": {
            "type": "integer",
            "title": "New Cases 7D",
            "default": 0
          },
          "visits_overdue": {
            "type": "integer",
            "title": "Visits Overdue",
            "default": 0
          },
          "returned_30d": {
            "type": "integer",
            "title": "Returned 30D",
            "default": 0
          },
          "attendance_compliance_pct": {
            "type": "number",
            "title": "Attendance Compliance Pct",
            "default": 0.0
          },
          "classes_missing_today": {
            "type": "integer",
            "title": "Classes Missing Today",
            "default": 0
          },
          "recent_escalations_count": {
            "type": "integer",
            "title": "Recent Escalations Count",
            "default": 0
          },
          "pending_help_requests_count": {
            "type": "integer",
            "title": "Pending Help Requests Count",
            "default": 0
          },
          "mentor_cases_active": {
            "type": "integer",
            "title": "Mentor Cases Active",
            "default": 0
          }
        },
        "type": "object",
        "title": "OverviewKPIs"
      },
...
      "RecentEscalationItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "school_name": {
            "type": "string",
            "title": "School Name"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "escalated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Escalated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ref",
          "student_name",
          "school_name",
          "level",
          "status",
          "escalated_at"
        ],
        "title": "RecentEscalationItem"
      },
...
      "ReportingGapItem": {
        "properties": {
          "class_name": {
            "type": "string",
            "title": "Class Name"
          },
          "school_name": {
            "type": "string",
            "title": "School Name"
          },
          "missing_dates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Dates"
          }
        },
        "type": "object",
        "required": [
          "class_name",
          "school_name"
        ],
        "title": "ReportingGapItem"
      },
...
      "SchoolCompareItem": {
        "properties": {
          "school_id": {
            "type": "string",
            "format": "uuid",
            "title": "School Id"
          },
          "school_name": {
            "type": "string",
            "title": "School Name"
          },
          "sector_name": {
            "type": "string",
            "title": "Sector Name"
          },
          "students_active": {
            "type": "integer",
            "title": "Students Active",
            "default": 0
          },
          "absent_today": {
            "type": "integer",
            "title": "Absent Today",
            "default": 0
          },
          "absence_rate_pct": {
            "type": "number",
            "title": "Absence Rate Pct",
            "default": 0.0
          },
          "open_cases": {
            "type": "integer",
            "title": "Open Cases",
            "default": 0
          },
          "returned_30d": {
            "type": "integer",
            "title": "Returned 30D",
            "default": 0
          },
          "attendance_compliance_pct": {
            "type": "number",
            "title": "Attendance Compliance Pct",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "school_id",
          "school_name",
          "sector_name"
        ],
        "title": "SchoolCompareItem"
      },
...
      "ScopeInfo": {
        "properties": {
          "level": {
            "type": "string",
            "title": "Level"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "level",
          "name"
        ],
        "title": "ScopeInfo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}